k8s: chart: rabbitmq: fix clustering
First, drop IP clustering and always set RABBITMQ_NODENAME to pod name, not IP.
We use StatefulSets for a reason. Otherwise, if the pod restarts (it can do so
at any time), RabbitMQ can suddenly change its storage directory and forget
everything in its queues.

Second, do not include loads of redundant information in the node name.
It is impossible to discover across namespaces, and resource names are unique in
a given namespace, hence there is no need to include anything else than the pod
name in the node name.